Add files via upload
Perform Basic Data Analysis Using Excel
Searching and sorting data organized in a spreadsheet are the most fundamental tasks associated with data analytics. Performing these operations on large volumes of data is not as simple as it sounds. This sprint will help to understand the different types and formats of data to begin with. Learners will use basic Excel features to search for data using lookup functions, sort data, and perform basis analysis to interpret what the data is telling them.