Add files via upload
Data Visualization using Excel
A picture is worth a thousand words, as the saying goes, and a visual in the form of a graph or chart speaks for the findings in the data that has been analyzed. In various situations, there are specific ways to present the results of analysis. In this sprint  I have learnt to choose basic visualization tools to create graphs and charts to visualize the insights obtained from data. These visualizations are usually centered around techniques in statistics.