Add files via upload
Measuring the Spread and Correlation of Data
Can a tweet by a presidential candidate affect the chances of winning or losing an election? The study involves observing cause and effect and is based on the correlation of data. New York City, one of the most expensive cities to live in the USA, has a range of rental properties ranging from $2,000 to $20,000 per month. What would the rental rate spread look like? Such a study involves measures of dispersion. By the end of this session, I have been able to analyze data using measures of dispersion and correlation.